requirements.txt
Pillow>=10.0.0
colorthief>=0.2.1
